Make `bundle install`work with clean state
There was no way to do a bundle install from a cloned version of this gem, because the requiring of the dependencies in the gemspec did not work.

The "rack/push-notification"` that was required at the top of the gemspec requires some more dependencies (rack, rack-contrib, sinatra, sequel), which are not present in case of a clean bundle install, because they are stated in the specification below.

So I moved the version declaration from a separate file directly to the gemspec: There is not really a need for having it in a version file, because the gemspec is the only file that needs the version information. It's somewhat of a bundler convention for having a separate file for that, but bundler does not enforce it.